Renamed hedconversion to be hedconverter
